Protected endpoints for departments admin features

+++ Mall-UI/src/app/departments/departments/departments.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { Department } from 'src/app/models/department';
+import { SecurityService } from 'src/app/security/security.service';
 import { DepartmentsService } from '../departments.service';
 
 @Component({
@@ -9,12 +11,19 @@ import { DepartmentsService } from '../departments.service';
 })
 export class DepartmentsComponent implements OnInit {
 
-  constructor(private departmentsService: DepartmentsService) { }
+  constructor(
+    private departmentsService: DepartmentsService,
+    private securityService: SecurityService,
+    private router: Router) { }
 
   departments: Department[] = []
 
   ngOnInit(): void {
-    this.getAll()
+    if(this.securityService.isLoggedIn()) {
+      this.getAll()
+    }else{
+      this.router.navigate(['/login'])
+    }
   }
 
   getAll(){
@@ -26,10 +35,14 @@ export class DepartmentsComponent implements OnInit {
   }
 
   delete(id: number){
-    this.departmentsService.delete(id)
-    .subscribe({
-      next: () => {this.reloadCurrentPage()}
-    })
+    if(this.securityService.isLoggedIn()) {
+      this.departmentsService.delete(id)
+      .subscribe({
+        next: () => {this.reloadCurrentPage()}
+      })
+    }else{
+      this.router.navigate(['/login'])
+    }
   }
 
   reloadCurrentPage(){
